fix serializatio tests
