+++ bittensor/serialization.py
@@ -1,6 +1,10 @@
+""" An interface for serializing and deserializing bittensor tensors"""
+import numpy as np
 import torch
+
+import bittensor
+import bittensor.utils.serialization_utils as serialization_utils
 from bittensor import bittensor_pb2
-from bittensor.serializers.pytorchpickle import PyTorchPickleSerializer
 
 class SerializationException (Exception):
     """ Raised during serialization """
@@ -23,8 +27,7 @@ class BittensorSerializerBase(object):
     various python tensor equivalents. i.e. torch.Tensor or tensorflow.Tensor
     """
 
-    @staticmethod
-    def serialize (tensor_obj: object, from_type: int) -> bittensor_pb2.Tensor:
+    def serialize (self, tensor_obj: object, modality: bittensor_pb2.Modality, from_type: int) -> bittensor_pb2.Tensor:
         """Serializes a torch object to bittensor_pb2.Tensor wire format.
 
         Args:
@@ -47,21 +50,20 @@ class BittensorSerializerBase(object):
         """
         # TODO (const): add deserialization types for torch -> tensorflow 
         if from_type == bittensor_pb2.TensorType.TORCH:
-            return BittensorSerializerBase.serialize_from_torch( tensor_obj )
+            return self.serialize_from_torch( torch_tensor = tensor_obj, modality = modality)
 
         elif from_type == bittensor_pb2.TensorType.NUMPY:
-            return BittensorSerializerBase.serialize_from_numpy( tensor_obj )
+            return self.serialize_from_numpy( numpy_tensor = tensor_obj, modality = modality)
 
         elif from_type == bittensor_pb2.TensorType.TENSORFLOW:
-            return BittensorSerializerBase.serialize_from_tensorflow( tensor_obj )
+            return self.serialize_from_tensorflow( tensorflow_tensor = tensor_obj, modality = modality)
 
         else:
             raise SerializationTypeNotImplementedException("Serialization from type {} not implemented.".format(from_type))
 
         raise NotImplementedError
 
-    @staticmethod
-    def deserialize (tensor_pb2: bittensor_pb2.Tensor, to_type: int) -> object:
+    def deserialize (self, tensor_pb2: bittensor_pb2.Tensor, to_type: int) -> object:
         """Serializes a torch object to bittensor_pb2.Tensor wire format.
 
         Args:
@@ -84,47 +86,36 @@ class BittensorSerializerBase(object):
         """
         # TODO (const): add deserialization types for torch -> tensorflow 
         if to_type == bittensor_pb2.TensorType.TORCH:
-            return BittensorSerializerBase.derserialize_to_torch( tensor_pb2 )
+            return self.deserialize_to_torch( tensor_pb2 )
 
         elif to_type == bittensor_pb2.TensorType.NUMPY:
-            return BittensorSerializerBase.derserialize_to_numpy( tensor_pb2 )
+            return self.deserialize_to_numpy( tensor_pb2 )
 
         elif to_type == bittensor_pb2.TensorType.TENSORFLOW:
-            return BittensorSerializerBase.derserialize_to_tensorflow( tensor_pb2 )
+            return self.deserialize_to_tensorflow( tensor_pb2 )
 
         else:
             raise SerializationTypeNotImplementedException("Deserialization to type {} not implemented.".format(to_type))
 
-    @staticmethod
-    def serialize_from_torch(torch_tensor: torch.Tensor) -> bittensor_pb2.Tensor:
+    def serialize_from_tensorflow(self, tensorflow_tensor: torch.Tensor, modality: bittensor_pb2.Modality) -> bittensor_pb2.Tensor:
         raise SerializationTypeNotImplementedException
 
-
-    @staticmethod
-    def derserialize_to_torch(tensor_pb2: bittensor_pb2.Tensor) -> torch.Tensor:
+    def serialize_from_torch(self, torch_tensor: torch.Tensor, modality: bittensor_pb2.Modality) -> bittensor_pb2.Tensor:
+        raise SerializationTypeNotImplementedException
+    
+    def serialize_from_numpy(self, numpy_tensor: torch.Tensor, modality: bittensor_pb2.Modality) -> bittensor_pb2.Tensor:
         raise SerializationTypeNotImplementedException
 
-    @staticmethod
-    def serialize_from_tensorflow(torch_tensor: torch.Tensor) -> bittensor_pb2.Tensor:
+    def deserialize_to_torch(self, tensor_pb2: bittensor_pb2.Tensor) -> torch.Tensor:
         raise SerializationTypeNotImplementedException
 
-
-    @staticmethod
-    def derserialize_to_tensorflow(tensor_pb2: bittensor_pb2.Tensor) -> torch.Tensor:
+    def deserialize_to_tensorflow(self, tensor_pb2: bittensor_pb2.Tensor) -> object:
         raise SerializationTypeNotImplementedException
 
-    @staticmethod
-    def serialize_from_numpy(torch_tensor: torch.Tensor) -> bittensor_pb2.Tensor:
+    def deserialize_to_numpy(self, tensor_pb2: bittensor_pb2.Tensor) -> object:
         raise SerializationTypeNotImplementedException
 
-
-    @staticmethod
-    def derserialize_to_numpy(tensor_pb2: bittensor_pb2.Tensor) -> torch.Tensor:
-        raise SerializationTypeNotImplementedException
-
-
-
-def get_serializer_for_type(serialzer_type: bittensor_pb2.Serializer) -> BittensorSerializerBase:
+def get_serializer ( serialzer_type: bittensor_pb2.Serializer ) -> BittensorSerializerBase:
     r"""Returns the correct serializer object for the passed Serializer enum. 
 
         Args:
@@ -141,7 +132,60 @@ def get_serializer_for_type(serialzer_type: bittensor_pb2.Serializer) -> Bittens
     """
 
     if serialzer_type == bittensor_pb2.Serializer.PICKLE:
-        return PyTorchPickleSerializer
+        return PyTorchPickleSerializer()
     else:
         raise NoSerializerForEnum("No known serialzier for proto type {}".format(serialzer_type))
+
+class PyTorchPickleSerializer( BittensorSerializerBase ):
+
+    def serialize_from_torch(self, torch_tensor: torch.Tensor, modality: bittensor_pb2.Modality) -> bittensor_pb2.Tensor:
+        """ Serializes a torch.Tensor to an bittensor Tensor proto.
+
+        Args:
+            torch_tensor (torch.Tensor): 
+                Torch tensor to serialize.
+
+            modality (bittensor_pb2.Modality): 
+                Datatype modality. i.e. TENSOR, TEXT, IMAGE
+
+        Returns:
+            bittensor_pb2.Tensor: 
+                The serialized torch tensor as bittensor_pb2.proto. 
+        """
+        if torch_tensor.requires_grad:
+            data_buffer = torch_tensor.cpu().detach().numpy().tobytes()
+        else:
+            data_buffer = torch_tensor.cpu().numpy().tobytes()
+        
+        dtype = serialization_utils.torch_dtype_to_bittensor_dtype(torch_tensor.dtype)
+        shape = list(torch_tensor.shape)
+        torch_proto = bittensor_pb2.Tensor(version = bittensor.__version__,
+                                    buffer = data_buffer,
+                                    shape = shape,
+                                    dtype = dtype,
+                                    serializer = bittensor_pb2.Serializer.PICKLE,
+                                    tensor_type = bittensor_pb2.TensorType.TORCH,
+                                    modality = modality,
+                                    requires_grad = torch_tensor.requires_grad)
+        return torch_proto
+
+
+    def deserialize_to_torch(self, torch_proto: bittensor_pb2.Tensor) -> torch.Tensor:
+        """Deserializes an bittensor_pb2.Tensor to a torch.Tensor object.
+
+        Args:
+            torch_proto (bittensor_pb2.Tensor): 
+                Proto containing torch tensor to derserialize.
+
+        Returns:
+            torch.Tensor: 
+                Deserialized torch tensor.
+        """
+        dtype = serialization_utils.bittensor_dtype_np_dtype(torch_proto.dtype)
+        array = np.frombuffer(torch_proto.buffer, dtype=np.dtype(dtype)).copy()
+        shape = tuple(torch_proto.shape)
+        tensor = torch.as_tensor(array).view(shape).requires_grad_(
+            torch_proto.requires_grad)
+        return tensor
+
