+++ tests/unit_tests/bittensor_tests/test_serialization.py
@@ -1,5 +1,6 @@
-from bittensor.serializer import DeserializationException, SerializationException
-from bittensor.serializer import PyTorchSerializer, torch_dtype_to_bittensor_dtype, bittensor_dtype_to_torch_dtype
+
+import bittensor.serialization as serialization
+import bittensor.utils.serialization_utils as serialization_utils
 from bittensor.config import Config
 from bittensor.subtensor.interface import Keypair
 from random import randrange
@@ -45,19 +46,30 @@ class TestSerialization(unittest.TestCase):
     def test_serialize(self):
         for _ in range(10):
             tensor_a = torch.rand([12, 23])
-            content = PyTorchSerializer.serialize(tensor_a, bittensor_pb2.Modality.TENSOR)
-            tensor_b = PyTorchSerializer.deserialize(content)
+            serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+            content = serializer.serialize(tensor_a, modality = bittensor_pb2.Modality.TENSOR, from_type = bittensor_pb2.TensorType.TORCH)
+            tensor_b = serializer.deserialize(content, to_type = bittensor_pb2.TensorType.TORCH)
             torch.all(torch.eq(tensor_a, tensor_b))
             
-    def test_serialize_modality(self):
+    def test_serialize_object_type_exception(self):
         # Let's grab some image data
         data = torchvision.datasets.MNIST(root = 'data/datasets/', train=True, download=True, transform=transforms.ToTensor())
         
         # Let's grab a random image, and try and de-serialize it incorrectly.
         image = data[randrange(len(data))][0]
 
-        with pytest.raises(SerializationException):
-            PyTorchSerializer.serialize(image, 11)
+        serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+        with pytest.raises(serialization.SerializationTypeNotImplementedException):
+            serializer.serialize(image, modality = bittensor_pb2.Modality.IMAGE, from_type = 11)
+
+    def test_deserialization_object_type_exception(self):
+        data = torch.rand([12, 23])
+        
+        serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+        tensor_message = serializer.serialize(data, modality = bittensor_pb2.Modality.TEXT, from_type = bittensor_pb2.TensorType.TORCH)
+
+        with pytest.raises(serialization.SerializationTypeNotImplementedException):
+            serializer.deserialize(tensor_message, to_type = 11)
     
     def test_serialize_deserialize_image(self):
         # Let's grab some image data
@@ -66,7 +78,8 @@ class TestSerialization(unittest.TestCase):
         # Let's grab a random image, and give it a crazy type to break the system
         image = data[randrange(len(data))][0]
 
-        serialized_image_tensor_message = PyTorchSerializer.serialize(image, bittensor_pb2.Modality.IMAGE)
+        serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+        serialized_image_tensor_message = serializer.serialize(image, modality = bittensor_pb2.Modality.IMAGE, from_type = bittensor_pb2.TensorType.TORCH)
         
         assert image.requires_grad == serialized_image_tensor_message.requires_grad
         assert list(image.shape) == serialized_image_tensor_message.shape
@@ -74,10 +87,10 @@ class TestSerialization(unittest.TestCase):
         assert serialized_image_tensor_message.dtype != bittensor_pb2.DataType.UNKNOWN
         assert serialized_image_tensor_message.buffer == image.cpu().numpy().tobytes()
 
-        deserialized_image_tensor_message = PyTorchSerializer.deserialize(serialized_image_tensor_message)
+        deserialized_image_tensor_message = serializer.deserialize(serialized_image_tensor_message, to_type = bittensor_pb2.TensorType.TORCH)
         assert serialized_image_tensor_message.requires_grad == deserialized_image_tensor_message.requires_grad
         assert serialized_image_tensor_message.shape == list(deserialized_image_tensor_message.shape)
-        assert torch_dtype_to_bittensor_dtype(deserialized_image_tensor_message.dtype) != bittensor_pb2.DataType.UNKNOWN
+        assert serialization_utils.torch_dtype_to_bittensor_dtype(deserialized_image_tensor_message.dtype) != bittensor_pb2.DataType.UNKNOWN
 
 
     def test_serialize_deserialize_text(self):
@@ -92,42 +105,38 @@ class TestSerialization(unittest.TestCase):
         data = torch.zeros((len(ts_list), max_l), dtype=torch.int64)
         for i, ts in enumerate(ts_list):
             data[i, 0:ts.size()[0]] = ts
-        
-        serialized_data_tensor_message = PyTorchSerializer.serialize(data, bittensor_pb2.Modality.TEXT)
+
+        serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+        serialized_data_tensor_message = serializer.serialize(data, modality = bittensor_pb2.Modality.TEXT, from_type = bittensor_pb2.TensorType.TORCH)
+       
         assert data.requires_grad == serialized_data_tensor_message.requires_grad
         assert list(data.shape) == serialized_data_tensor_message.shape
         assert serialized_data_tensor_message.modality == bittensor_pb2.Modality.TEXT
         assert serialized_data_tensor_message.dtype != bittensor_pb2.DataType.UNKNOWN
         assert serialized_data_tensor_message.buffer == data.cpu().numpy().tobytes()
 
-        deserialized_data_tensor_message = PyTorchSerializer.deserialize(serialized_data_tensor_message)
+        deserialized_data_tensor_message = serializer.deserialize(serialized_data_tensor_message, to_type = bittensor_pb2.TensorType.TORCH)
         assert serialized_data_tensor_message.requires_grad == deserialized_data_tensor_message.requires_grad
         assert serialized_data_tensor_message.shape == list(deserialized_data_tensor_message.shape)
-        assert torch_dtype_to_bittensor_dtype(deserialized_data_tensor_message.dtype) != bittensor_pb2.DataType.UNKNOWN
+        assert serialization_utils.torch_dtype_to_bittensor_dtype(deserialized_data_tensor_message.dtype) != bittensor_pb2.DataType.UNKNOWN
     
     def test_serialize_deserialize_tensor(self):
         data = torch.rand([12, 23])
 
-        serialized_tensor_message = PyTorchSerializer.serialize(data, bittensor_pb2.Modality.TENSOR)
+        serializer = serialization.get_serializer( serialzer_type = bittensor_pb2.Serializer.PICKLE )
+        serialized_tensor_message = serializer.serialize(data, modality = bittensor_pb2.Modality.TENSOR, from_type = bittensor_pb2.TensorType.TORCH)
+       
         assert data.requires_grad == serialized_tensor_message.requires_grad
         assert list(data.shape) == serialized_tensor_message.shape
         assert serialized_tensor_message.modality == bittensor_pb2.Modality.TENSOR
         assert serialized_tensor_message.dtype == bittensor_pb2.DataType.FLOAT32
         assert serialized_tensor_message.buffer == data.cpu().numpy().tobytes()
 
-        deserialized_tensor_message = PyTorchSerializer.deserialize(serialized_tensor_message)
+        deserialized_tensor_message = serializer.deserialize(serialized_tensor_message, to_type = bittensor_pb2.TensorType.TORCH)
         assert serialized_tensor_message.requires_grad == deserialized_tensor_message.requires_grad
         assert serialized_tensor_message.shape == list(deserialized_tensor_message.shape)
-        assert torch_dtype_to_bittensor_dtype(deserialized_tensor_message.dtype) == bittensor_pb2.DataType.FLOAT32
-
-    def test_deserialization_modality_failure(self):
-        data = torch.rand([12, 23])
-        tensor_message = PyTorchSerializer.serialize_tensor(data)
-        tensor_message.modality = 11
-
-        with pytest.raises(DeserializationException):
-            PyTorchSerializer.deserialize(tensor_message)
+        assert serialization_utils.torch_dtype_to_bittensor_dtype(deserialized_tensor_message.dtype) == bittensor_pb2.DataType.FLOAT32
     
     def test_bittensor_dtype_to_torch_dtype(self):
-        with pytest.raises(DeserializationException):
-            bittensor_dtype_to_torch_dtype(11)
+        with pytest.raises(serialization.DeserializationException):
+            serialization_utils.bittensor_dtype_to_torch_dtype(11)
